refactor: #13: move mmav to subfolder
