+++ rapiddb/src/db/mmav/mmav_database.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use std::io::{Read, Write};
 
-use crate::db::mmav::MMAV;
+use crate::db::mmav::mmav::MMAV;
 use crate::traits::IDatabase;
 use crate::types::AggregateFn;
