+++ rapiddb/src/db/mmav/mod.rs
@@ -1,0 +1,7 @@
+//! RapidDB Databases
+
+mod mmav;
+mod mmav_database;
+mod mmav_unit;
+
+pub use mmav_database::MMAVDatabase;
