+++ rapiddb/src/db/mod.rs
@@ -2,8 +2,8 @@
 
 mod database_test_factory;
 mod mmav;
-mod mmav_database;
-mod mmav_unit;
+// mod mmav_database;
+// mod mmav_unit;
 
 pub use database_test_factory::DatabaseTestFactory;
-pub use mmav_database::MMAVDatabase;
+pub use mmav::MMAVDatabase;
